move labelProperty to labels/
