+++ core/src/labels/labelProperty.cpp
