+++ core/src/labels/spriteLabel.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "labels/label.h"
-#include "style/labelProperty.h"
+#include "labels/labelProperty.h"
 #include "labels/labelSet.h"
 #include "labels/labelMesh.h"
 #include "style/pointStyle.h"
